+++ data/build.gradle.kts
@@ -26,4 +26,10 @@ dependencies {
     implementation(libs.koin.annotations)
     implementation(project(":shared"))
     ksp(libs.koin.compiler)
+
+    //Mockito, test coroutines, turbine
+    implementation(libs.mockito.kotlin)
+    implementation(libs.kotlin.coroutine.test)
+    implementation(libs.turbine)
+    testImplementation(libs.junit)
 }
