+++ useCases/src/test/java/com/airquality/usecases/FetchAirQualityUseCaseTest.kt
@@ -2,7 +2,7 @@ package com.airquality.usecases
 
 import app.cash.turbine.test
 import com.airquality.data.LocationAirQualityRepository
-import com.airquality.domain.datasource.HistoricForecastDataModel
+import com.airquality.domain.HomeDataModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.flowOf
@@ -18,16 +18,18 @@ import org.mockito.Mockito.mock
 import org.mockito.kotlin.whenever
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class GetForecastAirQualityUseCaseTest {
+class FetchAirQualityUseCaseTest {
 
-    private lateinit var useCase: GetForecastAirQualityUseCase
-    private val repository: LocationAirQualityRepository = mock()
+    private lateinit var useCase: FetchAirQualityUseCase
+
+    private val fakeRepository: LocationAirQualityRepository = mock()
+
     private val testDispatcher = StandardTestDispatcher()
 
     @Before
     fun setup() {
         Dispatchers.setMain(testDispatcher)
-        useCase = GetForecastAirQualityUseCase(repository)
+        useCase = FetchAirQualityUseCase(fakeRepository)
     }
 
     @After
@@ -36,10 +38,10 @@ class GetForecastAirQualityUseCaseTest {
     }
 
     @Test
-    fun `invoke should emit forecast data from repository`() = runTest {
+    fun `invoke should emit HomeDataModel from repository`() = runTest {
         // Arrange
-        val expected = HistoricForecastDataModel(city = "Barcelona")
-        whenever(repository.getForecastAirQuality()).thenReturn(flowOf(expected))
+        val expected = HomeDataModel(city = "Test City")
+        whenever(fakeRepository.getAirQuality()).thenReturn(flowOf(expected))
 
         // Act & Assert
         useCase().test {
